moved gff export function to dedicated script
